Fix tileconv paths for Linux & Max
The mod uses HANDLE_TILECONV, which in turn checks for the tileconv
executable in a subfolder named after the content of the %WEIDU_OS%
variable.

That variable holds 'osx' and 'unix', respectively, on Mac and Linux -
not 'macosx' and 'linux'. See:
http://www.weidu.org/~thebigg/README-WeiDU.html#WEIDU_OS

This commit renames those two subfolders accordingly.

I've tested it on Linux and confirmed that without this commit,
installation of the main component fails with:

    ERROR: "tileconv" does not exist at "ToA/tools/tileconv/unix/x86_64/tileconv"

whereas after this commit, the installation succeeds.